:dizzy:[update]Recommend取得をgameList,gameApiからfriendList,friendApiに移植、opponents取得処理修正
