+++ frontend/static_vol/css/style.css
@@ -788,7 +788,8 @@ input.has-input:invalid {
 
 /* no data */
 .blockFriends_friends p:only-child,
-.blockDashboardLog_listMatch p:only-child {
+.blockDashboardLog_listMatch p:only-child,
+.blockFriendRecommended_friends p:only-child {
   font-size: 1.8rem;
   margin-top: 1em;
   margin-bottom: 1em;
