+++ frontend/static_vol/js/modules/friendList.js
@@ -1,7 +1,7 @@
 'use strict';
 
-import { fetchFriendRequests, fetchFriends } from "./friendsApi.js";
-import { resetListenFriendList, resetListenFriendRequestList } from "./friendListener.js";
+import { fetchFriendRequests, fetchFriends, fetchRecommend } from "./friendsApi.js";
+import { addListenSendFriendRequest, resetListenFriendList, resetListenFriendRequestList } from "./friendListener.js";
 import { labels } from "./labels.js";
 import PageBase from "../components/PageBase.js";
 
@@ -96,4 +96,35 @@ const updateFriendRequestList = async (pageInstance) => {
     }
 }
 
-export { updateFriendsList, updateFriendRequestList }
+const updateRecommend = async (pageInstance) => {
+    console.log('updateRecommend');
+    try {
+        const RecommendedList = await fetchRecommend(false);
+        const RecommendedWrapper = document.querySelector('.blockFriendRecommended_friends');
+        if (!RecommendedList || RecommendedList.length === 0) {
+            RecommendedWrapper.innerHTML = `<p>${labels.friends.msgNoRecommended}</p>`
+        } else {
+            RecommendedWrapper.innerHTML = '';
+            RecommendedList.forEach(player => {
+                const avatar = player.avatar ? player.avatar : '/images/avatar_default.png';
+                const requestElement = `
+                    <section class="unitFriend">
+                        <header class="unitFriend_header">
+                            <h4 class="unitFriend_name">${player.username}</h4>
+                            <p class="unitFriend_thumb"><img src="${avatar}" alt="" width="100" height="100"></p>
+                        </header>
+                        <p class="unitFriendButton">
+                            <button type="button" class="unitFriendButton_friendRequest unitButton" data-username="${player.username}">${labels.friends.labelRequest}</button>
+                        </p>
+                    </section>
+                `;
+                RecommendedWrapper.innerHTML += requestElement;
+            });
+            addListenSendFriendRequest(pageInstance);
+        }
+    } catch (error) {
+        console.error('Failed to update recommended: ', error);
+    }
+}
+
+export { updateFriendsList, updateFriendRequestList, updateRecommend }
