+++ frontend/static_vol/js/modules/friendsApi.js
@@ -1,6 +1,6 @@
 import { getToken, refreshAccessToken } from './token.js';
 
-export const fetchFriends = async (isRefresh) => {
+const fetchFriends = async (isRefresh) => {
     const accessToken = getToken('accessToken');
     if (accessToken === null) {
         return Promise.resolve(null);
@@ -30,7 +30,7 @@ export const fetchFriends = async (isRefresh) => {
     }
 };
 
-export const fetchFriendRequests = async (isRefresh) => {
+const fetchFriendRequests = async (isRefresh) => {
     const accessToken = getToken('accessToken');
     if (accessToken === null) {
         return Promise.resolve(null);
@@ -59,3 +59,35 @@ export const fetchFriendRequests = async (isRefresh) => {
         console.error('Error fetching frined requests: ', error);
     }
 }
+
+const fetchRecommend = async (isRefresh) => {
+    const accessToken = getToken('accessToken');
+    if (accessToken === null) {
+        return Promise.resolve(null);
+    }
+    try {
+        const response = await fetch('/api/players/recommend/', {
+            headers: {
+                'Authorization': `Bearer ${accessToken}`
+            }
+        });
+        if (!response.ok) { //response.status=>403
+            if (!isRefresh) {
+                if (!await refreshAccessToken()) {
+                    throw new Error('fail refresh token');
+                }
+                return await fetchRecommend(true);
+            } else {
+                throw new Error('refreshed accessToken is invalid.');
+            }
+            throw new Error(`Failed to fetch recommended player: ${response.status}`);
+        }
+        const data = await response.json();
+        console.log('fetchRecommend API response: ', data);
+        return data;
+    } catch (error) {
+        console.error('Error fetching recommended player: ', error);
+    }
+}
+
+export { fetchFriends, fetchFriendRequests, fetchRecommend };
